+++ CHANGES.md
@@ -8,7 +8,9 @@
 - FIX: Fixing `xml.read` with cloud path as str
 - FIX: Fixing `ci.assert_dir_equal` if folders don't have files ordrered inside... (for some reason)
 - FIX: Add `distributed` in loggers removed by `ci.reduce_verbosity`
+- FIX: Use `sertit` LOGGER in `networks` module
 - DOC: Changing copyright from 2022 to 2023
+- CI: Better handling of logging display for pytest
 
 ## 1.21.2 (2022-12-14)